+++ Tests/StorageW1R3/Operation.swift
@@ -1,0 +1,32 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Foundation
+
+/// The type of storage operation performed during a benchmark step.
+public enum Operation: Sendable {
+  case resumable
+  case singleShot
+  case read(Int)
+  case delete
+
+  public var name: String {
+    switch self {
+    case .resumable: return "RESUMABLE"
+    case .singleShot: return "SINGLE_SHOT"
+    case .read(let index): return "READ[\(index)]"
+    case .delete: return "DELETE"
+    }
+  }
+}
